Change os_auth_url to use identity_endpoint in conf template
Current template ceilometer.conf.erb is using auth_uri for os_auth_url
which is not a correct approach. This auth_uri is dedicated to be used
by auth_uri under section [keystone_authtoken]. On the other hand, the
passed in identity_endpoint param is not used. Actually, by comparison
with other cookbooks, identity_endpoint should be used for keystone
identity service connection.

Change-Id: I50c631000efe9c82cd83723c1bf9a7c6ef54556c
Closes-bug: #1479399