Create attribute for setting options passed to ceilometer-upgrade
I ran into a problem similar to this one [1] where testing gnocchi
commands did not work properly. The key problem is the fact that
`--skip-gnocchi-resource-types` was included in the ceilometer-upgrade
command.  To allow an easy work around, I figure we can just create an
attribute which can add any option to ceilometer-upgrade so you can set
it to '' if you don't want it.

It might be debatable on whether to just default to run it with out
ceilometer-upgrade instead.

Some additional fixes include:

- Move the ceilometer-upgrade command to a new setup recipe where it
  makes more sense since it needs to be run after Ceilometer and Gnocchi
  are setup.
- Fix name of gnocchi-metricd_service for RHEL
- Add same apache restart fixes that we implemented in other cookbooks.
  This is needed so that you can properly run ceilometer-upgrade with
  gnocchi enabled.

[1] https://bugs.launchpad.net/openstack-ansible/+bug/1737096

Change-Id: I619ef044b8cb254b23e0c7bc674c46d5dd7e0076
Signed-off-by: Lance Albertson <lance@osuosl.org>