Manage /etc/ceilometer/polling.yaml
This adds management of /etc/ceilometer/polling.yaml which defines which metrics
to poll and how often. This pulls in the default configuration which the same
for both RHEL and Ubuntu and adds two new attributes for configuring the file.

In addition, restart various services related to the configuration file.

Change-Id: I69ed0a8933a971d9cac20dbd4133197a7e4274af
Signed-off-by: Lance Albertson <lance@osuosl.org>